Revert "change to 64"
This reverts commit f75e0a46e95aece442dab190fb652e0cc86da1bb.

+++ QuICT/core/gate/gate_unit_test.py
@@ -23,7 +23,7 @@ def test_permMulDetail():
             ControlPermMulDetail([2, 5]).inverse() | circuit
             X | circuit
             unitary = SyntheticalUnitary.run(circuit)
-            if (abs(abs(unitary - np.identity((1 << i), dtype=np.complex64))) > 1e-10).any():
+            if (abs(abs(unitary - np.identity((1 << i), dtype=np.complex128))) > 1e-10).any():
                 assert 0
     assert 1
 
@@ -43,7 +43,7 @@ def test_CCRz():
             # print(amplitude)
             X | circuit
             unitary = SyntheticalUnitary.run(circuit)
-            if (abs(abs(unitary - np.identity((1 << i), dtype=np.complex64))) > 1e-10).any():
+            if (abs(abs(unitary - np.identity((1 << i), dtype=np.complex128))) > 1e-10).any():
                 assert 0
             assert 1
     assert 1
@@ -77,7 +77,7 @@ def test_fSim():
                 print(amplitude)
                 assert 0
             unitary = SyntheticalUnitary.run(circuit)
-            if (abs(abs(unitary - np.identity((1 << i), dtype=np.complex64))) > 1e-10).any():
+            if (abs(abs(unitary - np.identity((1 << i), dtype=np.complex128))) > 1e-10).any():
                 # print(unitary)
                 assert 0
     assert 1
@@ -101,7 +101,7 @@ def test_Rxx():
                 print(amplitude)
                 assert 0
             unitary = SyntheticalUnitary.run(circuit)
-            if (abs(abs(unitary - np.identity((1 << i), dtype=np.complex64))) > 1e-10).any():
+            if (abs(abs(unitary - np.identity((1 << i), dtype=np.complex128))) > 1e-10).any():
                 # print(unitary)
                 assert 0
     assert 1
@@ -125,7 +125,7 @@ def test_Ryy():
                 print(amplitude)
                 assert 0
             unitary = SyntheticalUnitary.run(circuit)
-            if (abs(abs(unitary - np.identity((1 << i), dtype=np.complex64))) > 1e-10).any():
+            if (abs(abs(unitary - np.identity((1 << i), dtype=np.complex128))) > 1e-10).any():
                 # print(unitary)
                 assert 0
     assert 1
@@ -149,7 +149,7 @@ def test_Rzz():
                 print(amplitude)
                 assert 0
             unitary = SyntheticalUnitary.run(circuit)
-            if (abs(abs(unitary - np.identity((1 << i), dtype=np.complex64))) > 1e-10).any():
+            if (abs(abs(unitary - np.identity((1 << i), dtype=np.complex128))) > 1e-10).any():
                 # print(unitary)
                 assert 0
     assert 1
